feat: extract utility functions from AppView.tsx to src/lib/ modules
Move ANSI rendering, formatters, and log/data utilities out of AppView.tsx
into dedicated lib files (ansi.ts, formatters.ts, log.ts). Re-export all
previously-exported symbols from AppView for backward compatibility with tests.

Aloop-Agent: build
Aloop-Iteration: 3
Aloop-Session: orchestrator-20260321-172932-issue-157-20260323-094646
